requirements.txt
Flask==0.12.2
python-dotenv==0.6.4
mysqlclient==1.3.12
# Flask extensions
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
alembic==0.9.3

# Linter
flake8==2.4.0

# faker - to generate fake data(used in seeder)
Faker==0.7.18


# google token check
oauth2client==4.1.2

# requests
requests==2.18.4

#handle twitter oauth
oauth2==1.9.0.post1
Flask-Uploads==0.2.1
Flask-SocketIO==2.9.2
Flask-Mail==0.9.1
python_resize_image==1.1.11
flask-cors==3.0.3

#handle decimal in json
Simplejson==3.11.1

paypalrestsdk==1.13.1